simplified component testing
