+++ .vscode/extensions.json
@@ -1,3 +1,3 @@
 {
-  "recommendations": ["ms-playwright.playwright"]
+  "recommendations": ["ms-playwright.playwright", "esbenp.prettier-vscode"]
 }
