+++ apps/component-tests/.prettierignore
@@ -1,0 +1,6 @@
+# Files Prettier should not format
+**/*.log
+**/.DS_Store
+*.
+dist
+node_modules
