+++ apps/component-tests/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "componentTests"
+}
