+++ apps/component-tests/postcss.config.cjs
@@ -1,0 +1,11 @@
+const path = require('path');
+
+module.exports = {
+  plugins: {
+    'tailwindcss/nesting': 'postcss-nested',
+    tailwindcss: {
+      config: path.join(__dirname, 'tailwind.config.cjs'),
+    },
+    autoprefixer: {},
+  },
+};
