+++ apps/component-tests/src/entry.preview.tsx
@@ -1,0 +1,21 @@
+/*
+ * WHAT IS THIS FILE?
+ *
+ * It's the bundle entry point for `npm run preview`.
+ * That is, serving your app built in production mode.
+ *
+ * Feel free to modify this file, but don't remove it!
+ *
+ * Learn more about Vite's preview command:
+ * - https://vitejs.dev/config/preview-options.html#preview-options
+ *
+ */
+import { createQwikCity } from '@builder.io/qwik-city/middleware/node';
+// The `render` import must stay after `qwikCityPlan` import: https://github.com/BuilderIO/qwik/issues/3634
+import qwikCityPlan from '@qwik-city-plan';
+import render from './entry.ssr';
+
+/**
+ * The default export is the QwikCity adaptor used by Vite preview.
+ */
+export default createQwikCity({ render, qwikCityPlan });
