+++ apps/component-tests/src/global.css
@@ -1,0 +1,3 @@
+@tailwind components;
+@tailwind base;
+@tailwind utilities;
