+++ apps/component-tests/src/routes/[kit]/[component]/[example]/index@test.tsx
@@ -1,0 +1,6 @@
+import { component$ } from '@builder.io/qwik';
+import { ShowcaseTest } from '../../../../components/showcase-test/showcase-test';
+
+export default component$(() => {
+  return <ShowcaseTest />;
+});
