+++ apps/component-tests/vite.config.ts
@@ -1,0 +1,34 @@
+import { qwikVite } from '@builder.io/qwik/optimizer';
+import { qwikCity } from '@builder.io/qwik-city/vite';
+import { defineConfig } from 'vite';
+import tsconfigPaths from 'vite-tsconfig-paths';
+import { qwikNxVite } from 'qwik-nx/plugins';
+
+export default defineConfig({
+  cacheDir: '../../node_modules/.vite/apps/component-tests',
+  plugins: [
+    qwikNxVite(),
+    qwikCity(),
+    qwikVite({
+      client: {
+        outDir: '../../dist/apps/component-tests/client',
+      },
+      ssr: {
+        outDir: '../../dist/apps/component-tests/server',
+      },
+      tsconfigFileNames: ['tsconfig.app.json'],
+    }),
+    tsconfigPaths({ root: '../../' }),
+  ],
+  server: {
+    fs: {
+      // Allow serving files from the project root
+      allow: ['../../'],
+    },
+  },
+  preview: {
+    headers: {
+      'Cache-Control': 'public, max-age=600',
+    },
+  },
+});
