+++ apps/website/project.json
@@ -80,13 +80,6 @@
       "options": {
         "command": "npx nx run website:build:production"
       }
-    },
-    "e2e": {
-      "executor": "@nx/playwright:playwright",
-      "outputs": ["{workspaceRoot}/dist/.playwright/apps/website"],
-      "options": {
-        "config": "apps/website/playwright.config.ts"
-      }
     }
   },
   "tags": []
