+++ apps/website/src/routes/docs/headless/select/[example]/index@test.tsx
@@ -1,0 +1,6 @@
+import { component$ } from '@builder.io/qwik';
+import { ShowcaseTest } from '~/components/showcase-test/showcase-test';
+
+export default component$(() => {
+  return <ShowcaseTest />;
+});
