+++ apps/website/src/routes/layout-test.tsx
@@ -1,0 +1,9 @@
+import { component$, Slot } from '@builder.io/qwik';
+
+export default component$(() => {
+  return (
+    <>
+      <Slot />
+    </>
+  );
+});
