+++ apps/website/tsconfig.json
@@ -26,5 +26,5 @@
       "path": "./tsconfig.spec.json"
     }
   ],
-  "include": ["src", "src/routes/docs/headless/select/select.spec.ts"]
+  "include": ["src", "../../packages/kit-headless/src/components/select/select.test.ts"]
 }
