+++ package.json
@@ -15,7 +15,6 @@
     "change": "changeset",
     "changeset.version": "changeset version && pnpm install --no-frozen-lockfile && git add --all",
     "dev": "nx serve website",
-    "test.docs": "nx e2e website",
     "format.fix": "pretty-quick --staged",
     "link.dist": "cd dist/packages/kit-headless && pnpm link --global",
     "lint": "nx affected:lint",
@@ -32,6 +31,7 @@
     "story.headless": "nx storybook headless",
     "test.cli": "nx test cli",
     "test.headless": "nx component-test headless",
+    "test.pw.headless": "nx e2e headless",
     "test.headless.ci": "nx component-test-ci headless",
     "test.utils": "nx test utils"
   },
@@ -215,5 +215,6 @@
   ],
   "nx": {
     "includedScripts": []
-  }
+  },
+  "dependencies": {}
 }
